requirements.txt
fastapi
uvicorn
opencv-python
imutils
scipy
streamlit
requests
pillow
dlib
